Develop (\
* [USD-59] feat: 현재 재생중인 영상 변경 API 구성 및 SSE 설정 완료

* [USD-55] feat: 재생중인 영상 삭제 불가능하게 변경

* [USD-67] feat: 유튜브 링크가 이상할 때 에러 던짐

+++ src/main/java/com/uhsadong/ddtube/domain/service/VideoCommandService.java
@@ -64,6 +64,10 @@ public class VideoCommandService {
         if (!(video.getUser().getId().equals(user.getId()) || user.isAdmin())) {
             throw new GeneralException(ErrorStatus._VIDEO_DELETE_PERMISSION_DENIED);
         }
+        if (playlist.getNowPlayVideo() != null && video.getId()
+            .equals(playlist.getNowPlayVideo().getId())) {
+            throw new GeneralException(ErrorStatus._CANNOT_DELETE_NOW_PLAY_VIDEO);
+        }
 
         videoRepository.delete(video);
         sseEmitters.sendVideoEventToClients(playlistCode, video, SseStatus.DELETE);
